[bale]Add the types of art films.

+++ src/main/java/rentalstore/HtmlStatement.java
@@ -13,7 +13,7 @@ public class HtmlStatement extends Statement{
                 + customer.getName()
                 + "</EM></H1><P>\n";
     }
-    protected  String getFootString(double totalAmount, int frequentRenterPoints){
+    protected  String getFootString(double totalAmount, double frequentRenterPoints){
         return "<P>You owe<EM>"
                 + String.valueOf(totalAmount)
                 + "</EM><p>\n"
