+++ src/main/java/rentalstore/MovieType.java
@@ -2,7 +2,7 @@ package rentalstore;
 
 public abstract class MovieType {
     protected abstract double getCharge(int dayRented);
-    public int getFrequentRenterPoints(int dayRented) {
+    public double getFrequentRenterPoints(int dayRented) {
         return 1;
     }
 }
